Created Simple World Generator
Basic proof of concept of the world generation and noise generation. The
scene now has a red button in the top corner, which, when pressed, calls
the new demo generation script that simply maps a noise map to walkable
and non-walkable tiles. The values of local variables in the
NoiseGenerator class and the WorldGenDemo class have been tested, and
are appropriate for the demo.